* Redux for state management
We were passing around state explicitly which isn't necessarily bad; but,
as the app grew in terms of features the state was growing too and it was
becoming harder to see the wood for the trees.

There's now one state tree managed by Redux.

* Internationalisation

The app is internationalised (I18N) and localised (L10N) for English only.

It's annoying donkey work to do I18N and L10N after the fact; if you just
get such support int your codebase early then it's not really any more
effort to use the I18N framework.

* Draft.js for editing the Interactive Fiction (IF)

Still working through this one :thinking It _feels_ like it could be
suitable for our needs but we have to really kick the tyres on it and
solicit feedback from actual users.

* Uppy for file uploads

https://uppy.io/